requirements.txt
Flask
Flask-SQLAlchemy
sqlalchemy
pytest
pymysql
Pillow